requirements.txt
django
django-cors-headers
djangorestframework
requests
psycopg2
django-debug-toolbar
djangorestframework-simplejwt
